Add q1 code
Note that section be takes a lot of time to complete. The best values are found in a comment there, but I have fixed a bug, so not sure they are the same now. I am running the code on my machine to get those values, and I'll update you with them once this completes.
Thank you!!